strong_hash

Trait StrongHash

pub trait StrongHash {
    // Required method
    fn strong_hash<H: Hasher>(&self, state: &mut H);
}

StrongHash`` is a trait that is notionally similar to std::hash::Hash`, but carries the implicit expectation that the hash that will be produced should be as perturbed as possible.

StrongHash must be implemented such that - when combined with a crypto hasher - the hash value is suitable to use for equality checks. That means fields must not be omitted and pre-computed hash values cannot be used if the hasher used for them is weak.

StrongHash can be derived on enums and structs if all of their members implement StrongHash. For example:

#[derive(StrongHash)]
struct MyStruct {
    a: u32,
    b: String,
    c: Vec<u8>,
}

StrongHash can also be implemented manually similar to std::hash::Hash. For example:

struct MyStruct {
    a: u32,
    b: String,
    c: Vec<u8>,
}

impl StrongHash for MyStruct {
    fn strong_hash<H: StrongHasher>(&self, state: &mut H) {
        self.a.strong_hash(state);
        self.b.strong_hash(state);
        self.c.strong_hash(state);
    }
}

Required Methods

fn strong_hash<H: Hasher>(&self, state: &mut H)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.